refactor: implement useRecaptcha composable and refactor rendering logic
