+++ app/plugins/recapcha.client.ts
@@ -11,40 +11,4 @@ export default defineNuxtPlugin(() => {
     script.onerror = () => console.error('Failed to load reCAPTCHA script');
     document.head.appendChild(script);
   }
-
-  const renderRecaptcha = ({
-    elementId,
-    callback,
-    expiredCallback,
-  }: {
-    elementId: string;
-    callback: (token: string) => void;
-    expiredCallback?: () => void;
-  }) => {
-    const siteKey = useRuntimeConfig().public.siteKey || '';
-    const theme = document.documentElement.classList.contains('dark') ? 'dark' : 'light';
-    const render = () => {
-      window.grecaptcha.render(elementId, {
-        sitekey: siteKey,
-        callback,
-        'expired-callback': expiredCallback,
-        theme,
-        hl: document.documentElement.lang || 'en',
-      });
-    };
-
-    if (window.grecaptcha) {
-      render();
-    } else {
-      window.addEventListener('recaptcha-script-loaded', render, { once: true });
-    }
-  };
-
-  return {
-    provide: {
-      recaptcha: {
-        render: renderRecaptcha,
-      },
-    },
-  };
 });
